Add user authentication
This commit adds authentication functionality to the page. Now users
will have to login to access their tasks, and are forbidden from
accessing tasks that are not their own.

+++ todo/tasks/views.py
@@ -1,14 +1,21 @@
 from django.shortcuts import render, redirect
+from django.urls import reverse
 from . import models, forms
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 
+@login_required
 def index(request):
     if request.method == "POST":
         form = forms.TaskForm(request.POST)
         if form.is_valid():
-            form.save()
-        return redirect('/')
+            task = form.save(commit=False)
+            task.user = request.user
+            task.save()
 
-    tasks = models.Task.objects.all()
+        return redirect(reverse('list'))
+
+    tasks = models.Task.objects.filter(user=request.user)
     form = forms.TaskForm
     context = {
         'tasks': tasks,
@@ -17,15 +24,18 @@ def index(request):
     return render(request, 'tasks/list.html', context)
 
 # pk comes from the url pattern
+@login_required
 def update_task(request, pk):
     task = models.Task.objects.get(id=pk)
+    if task.user != request.user:
+        return HttpResponseForbidden("The user does not have permission to access this task.")
 
     if request.method == "POST":
         form = forms.TaskForm(request.POST, instance=task)
 
         if form.is_valid():
             form.save()
-            return redirect('/')
+            return redirect(reverse('list'))
 
     form = forms.TaskForm(instance=task)
     context = {
@@ -34,12 +44,15 @@ def update_task(request, pk):
 
     return render(request, "tasks/update_task.html", context)
 
+@login_required
 def delete_task(request, pk):
     task = models.Task.objects.get(id=pk)
+    if task.user != request.user:
+        return HttpResponseForbidden("The user does not have permission to access this task.")
 
     if request.method == "POST":
         task.delete()
-        return redirect('/')
+        return redirect(reverse('list'))
 
     context = {
         "task": task,
